+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Aurarium/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Aurarium/releases)
 
-- Added: Currencies for the expansion 'Midnight' [retail]
+- Adapted to the latest version of Arcane Wizard: Library to ensure full compatibility